Twitter Troll analysis
NBC posted a dataset for posts and users believed to be originating from Russian troll accounts. There are over 200,000 tweets, primarily dating back to the 2016 election, contained within the dataset. This code extracts the hashtags used in each of the tweets and in to a set of hashtags, and from there extracts which hashtags the troll accounts were using the most. The output is a dictionary containing every single hash tag and how often it was used, and also two graphs demonstrating how often each of the presidential candidates were mentioned singularly and alongside one another.